+++ src/ModuleConfig/PathFinder/BasePathFinder.php
@@ -3,10 +3,13 @@ namespace Qobo\Utils\ModuleConfig\PathFinder;
 
 use Cake\Core\Configure;
 use InvalidArgumentException;
+use Qobo\Utils\ErrorTrait;
 use Qobo\Utils\Utility;
 
 abstract class BasePathFinder implements PathFinderInterface
 {
+    use ErrorTrait;
+
     /**
      * CakePHP configuration key with base path
      *
@@ -25,16 +28,6 @@ abstract class BasePathFinder implements PathFinderInterface
     protected $fileName;
 
     /**
-     * @var array $errors List of errors from last find
-     */
-    protected $errors = [];
-
-    /**
-     * @var array $errors List of warnings from last find
-     */
-    protected $warnings = [];
-
-    /**
      * Find path
      *
      * @param string $module Module to look for files in
@@ -45,11 +38,11 @@ abstract class BasePathFinder implements PathFinderInterface
     public function find($module, $path = null, $validate = true)
     {
         if (empty($module)) {
-            $this->fail("Module is not specified");
+            $this->fail(new InvalidArgumentException("Module is not specified"));
         }
 
         if (!is_string($module)) {
-            $this->fail("Module name is not a string");
+            $this->fail(new InvalidArgumentException("Module name is not a string"));
         }
 
         if (empty($path)) {
@@ -57,7 +50,7 @@ abstract class BasePathFinder implements PathFinderInterface
         }
 
         if (!is_string($path)) {
-            $this->fail("Path is not a string");
+            $this->fail(new InvalidArgumentException("Path is not a string"));
         }
 
         $result = '';
@@ -77,41 +70,5 @@ abstract class BasePathFinder implements PathFinderInterface
         }
 
         return $result;
-    }
-
-    /**
-     * Fail execution with a given error
-     *
-     * * Adds error to the list of errors
-     * * Throws an exception with the error message
-     *
-     * @throws \InvalidArgumentException
-     * @param string $message Error message
-     * @return void
-     */
-    protected function fail($message)
-    {
-        $this->errors[] = $message;
-        throw new InvalidArgumentException($message);
-    }
-
-    /**
-     * Get finder errors
-     *
-     * @return array List of errors from last find
-     */
-    public function getErrors()
-    {
-        return $this->errors;
-    }
-
-    /**
-     * Get finder warnings
-     *
-     * @return array List of warnings from last find
-     */
-    public function getWarnings()
-    {
-        return $this->warnings;
     }
 }
